Move *nix specific implementation to separate module
