+++ src/hostcalls/mod.rs
@@ -1,7 +1,4 @@
-//! Hostcalls that implement
-//! [WASI](https://github.com/CraneStation/wasmtime-wasi/blob/wasi/docs/WASI-overview.md).
 mod fs;
-mod fs_helpers;
 mod misc;
 mod sock;
