+++ src/lib.rs
@@ -20,11 +20,12 @@
     )
 )]
 
-pub mod ctx;
-pub mod fdentry;
+mod ctx;
+mod sys;
+
+pub mod memory;
 pub mod host;
 pub mod hostcalls;
-pub mod memory;
 pub mod wasm32;
 
 pub use ctx::{WasiCtx, WasiCtxBuilder};
